refactor icon
